backend/requirements.txt
flask==2.2.3
flask-restful==0.3.9
flask-jwt-extended==4.4.4
pymongo==4.3.3
flask-cors==3.0.10
python-dotenv==1.0.0
werkzeug==2.2.3
kubernetes

# Database connectors
mysql-connector-python==8.0.33
psycopg2-binary==2.9.6
neo4j==5.14.1

# AWS SDK
boto3==1.28.38
botocore==1.31.38

# Utilities
requests==2.31.0
#pyjwt==2.6.0

# GitHub and IAC scanning
GitPython==3.1.40
checkov
#cyclonedx-bom<4.0.0
#cryptography==41.0.3
python-dateutil==2.8.2
pytz==2023.3
